Refactor: 이미지 서비스 코드 리팩토링
- 파사드 패턴 적용
- 직접 Image Upload, Delete 기능을 사용할 필요가 없게 됨에 따라 파사드 패턴으로 분리
- 파사드 패턴에서는 Spring event 등록과 ImagePath에 대한 Return 기능만 제공
